+++ R/listdial.R
@@ -20,7 +20,7 @@
 #'
 #' Creates a new Dialer and binds it to a Socket.
 #'
-#' @param socket a Socket or nano object.
+#' @param socket a Socket.
 #' @param url [default 'inproc://nanonext'] a URL to dial, specifying the
 #'     transport and address as a character string e.g. 'inproc://anyvalue' or
 #'     'tcp://127.0.0.1:5555' (see \link{transports}).
@@ -30,18 +30,12 @@
 #'
 #' @return Invisibly, an integer exit code (zero on success). A new Dialer
 #'     (object of class 'nanoDialer' and 'nano') is created and bound to the
-#'     Socket or nano object if successful.
+#'     Socket if successful.
 #'
 #' @details To view all Dialers bound to a socket use \code{$dialer} on the
 #'     socket, which returns a list of Dialer objects. To access any individual
 #'     Dialer (e.g. to set options on it), index into the list e.g.
 #'     \code{$dialer[[1]]} to return the first Dialer.
-#'
-#'     This function may be used to bind a new Dialer to a Socket, or else a
-#'     nano object. If called on a nano object, the dialer is attached to the
-#'     object rather than the socket for ease of access, e.g. \code{$dialer[[1]]}
-#'     rather than \code{$socket$dialer[[1]]}, but is otherwise equivalent to
-#'     calling \code{dial()} on the object's socket directly.
 #'
 #'     A Dialer is an external pointer to a dialer object, which creates a
 #'     single outgoing connection at a time. If the connection is broken, or
@@ -76,9 +70,9 @@
 #' close(socket)
 #'
 #' nano <- nano("bus")
-#' dial(nano, url = "tcp://127.0.0.1:6546", autostart = FALSE)
+#' nano$dial(url = "tcp://127.0.0.1:6546", autostart = FALSE)
 #' nano$dialer
-#' start(nano$dialer[[1]])
+#' nano$dialer_start()
 #' nano$dialer
 #' close(nano$dialer[[1]])
 #' nano$close()
@@ -91,21 +85,7 @@ dial <- function(socket,
 
   res <- if (autostart) .Call(rnng_dial, socket, url) else .Call(rnng_dialer_create, socket, url)
   is.integer(res) && return(invisible(res))
-
-  if (is.environment(socket)) {
-    socket[["dialer"]] <- c(.subset2(socket, "dialer"), res)
-    socket[["dialer_setopt"]] <- function(type = c("bool", "int", "ms", "size",
-                                                   "string", "uint64"),
-                                          opt,
-                                          value) invisible(lapply(.subset2(socket, "dialer"),
-                                                                  setopt,
-                                                                  type = type,
-                                                                  opt = opt,
-                                                                  value = value))
-  } else {
-    attr(socket, "dialer") <- c(attr(socket, "dialer"), res)
-  }
-
+  attr(socket, "dialer") <- c(attr(socket, "dialer"), res)
   invisible(0L)
 
 }
@@ -114,7 +94,7 @@ dial <- function(socket,
 #'
 #' Creates a new Listener and binds it to a Socket.
 #'
-#' @param socket a Socket or nano object.
+#' @param socket a Socket.
 #' @param url [default 'inproc://nanonext'] a URL to dial or listen at, specifying
 #'     the transport and address as a character string e.g. 'inproc://anyvalue'
 #'     or 'tcp://127.0.0.1:5555' (see \link{transports}).
@@ -124,18 +104,12 @@ dial <- function(socket,
 #'
 #' @return Invisibly, an integer exit code (zero on success). A new Listener
 #'     (object of class 'nanoListener' and 'nano') is created and bound to the
-#'     Socket or nano object if successful.
+#'     Socket if successful.
 #'
 #' @details To view all Listeners bound to a socket use \code{$listener} on the
 #'     socket, which returns a list of Listener objects. To access any individual
 #'     Listener (e.g. to set options on it), index into the list e.g.
 #'     \code{$listener[[1]]} to return the first Listener.
-#'
-#'     This function may be used to bind a new Listener to a Socket, or else a
-#'     nano object. If called on a nano object, the listener is attached to the
-#'     object rather than the socket for ease of access, e.g. \code{$listener[[1]]}
-#'     rather than \code{$socket$listener[[1]]}, but is otherwise equivalent to
-#'     calling \code{listen()} on the object's socket directly.
 #'
 #'     A listener is an external pointer to a listener object, which accepts
 #'     incoming connections. A given listener object may have many connections
@@ -170,9 +144,9 @@ dial <- function(socket,
 #' close(socket)
 #'
 #' nano <- nano("bus")
-#' listen(nano, url = "tcp://127.0.0.1:6548", autostart = FALSE)
+#' nano$listen(url = "tcp://127.0.0.1:6548", autostart = FALSE)
 #' nano$listener
-#' start(nano$listener[[1]])
+#' nano$listener_start()
 #' nano$listener
 #' close(nano$listener[[1]])
 #' nano$close()
@@ -185,21 +159,7 @@ listen <- function(socket,
 
   res <- if (autostart) .Call(rnng_listen, socket, url) else .Call(rnng_listener_create, socket, url)
   is.integer(res) && return(invisible(res))
-
-  if (is.environment(socket)) {
-    socket[["listener"]] <- c(.subset2(socket, "listener"), res)
-    socket[["listener_setopt"]] <- function(type = c("bool", "int", "ms", "size",
-                                                   "string", "uint64"),
-                                            opt,
-                                            value) invisible(lapply(.subset2(socket, "listener"),
-                                                                    setopt,
-                                                                    type = type,
-                                                                    opt = opt,
-                                                                    value = value))
-  } else {
-    attr(socket, "listener") <- c(attr(socket, "listener"), res)
-  }
-
+  attr(socket, "listener") <- c(attr(socket, "listener"), res)
   invisible(0L)
 
 }